renaming DateTimeView -> DateView
