+++ packages/components/src/DateTimePickers/views/DateView/DateView.scss
