+++ packages/components/src/DateTimePickers/views/DateView/index.js
@@ -1,5 +1,5 @@
 import { withTranslation } from 'react-i18next';
-import DateTimeView from './DateTimeView.component';
+import DateTimeView from './DateView.component';
 import I18N_DOMAIN_COMPONENTS from '../../../constants';
 
 export default withTranslation(I18N_DOMAIN_COMPONENTS)(DateTimeView);
